@@ -7,6 +7,8 @@
   "types": "index.d.ts",
   "dependencies": {
     "@babel/runtime": "^7.23.2",
+    "@noble/ciphers": "^0.5.1",
+    "@noble/curves": "^1.4.0",
     "@noble/post-quantum": "../noble-post-quantum",
     "@panva/hkdf": "^1.0.4",
     "@stablelib/base64": "^0.7.2",
@@ -33,6 +35,7 @@
     "@babel/preset-typescript": "^7.23.2",
     "@types/faker": "^4.1.5",
     "@types/jest": "^24.0.15",
+    "@types/node": "^20.11.30",
     "argon2": "^0.30.3",
     "argon2-browser": "^1.18.0",
     "events": "^3.3.0",
@@ -43,7 +46,7 @@
     "process": "^0.11.10",
     "stream-browserify": "^3.0.0",
     "ts-jest": "^24.0.2",
-    "typescript": "^3.5.2"
+    "typescript": "^5.4.2"
   },
   "scripts": {
     "watch": "./node_modules/.bin/babel src --watch -d dist --extensions \".ts,.js\" --source-maps inline",
